# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "API Operaçao Container",
        "description": "Documentação da API"
    },
    "servers": [
        {
            "url": "",
            "description": "Servidor local"
        },
        {
            "url": "https://api-operacao-container-production.up.railway.app",
            "description": "Servidor de Produção"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token de autenticação Bearer. Coloque o token JWT gerado na caixa abaixo no formato `Bearer <token>`."
            }
        },
        "schemas": {
            "usuarioLogin": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Nome de usuário",
                        "example": "SunnyAura.suporte"
                    },
                    "senha": {
                        "type": "string",
                        "description": "Senha do usuário",
                        "example": "123456"
                    }
                }
            },
            "usuarioCreate": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "Nome de usuário",
                        "example": "SunnyAura.novo"
                    },
                    "senha": {
                        "type": "string",
                        "description": "Senha do usuário",
                        "example": "minhasenha"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email do usuário",
                        "example": "novo@email.com"
                    }
                }
            },
            "usuarioAutenticado": {
                "type": "object",
                "properties": {
                    "usuario": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "username": {
                                "type": "string",
                                "example": "SunnyAura.suporte"
                            },
                            "email": {
                                "type": "string",
                                "example": "sunnyaura@gmail.com"
                            },
                            "cliente": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "razaoSocial": {
                                        "type": "string",
                                        "example": "Sunny Aura"
                                    },
                                    "nomeFantasia": {
                                        "type": "string",
                                        "example": "Sunny Aura"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "46.095.984/0001-32"
                                    },
                                    "filiais": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nomeFilial": {
                                                    "type": "string"
                                                },
                                                "endereco": {
                                                    "type": "string"
                                                },
                                                "cidade": {
                                                    "type": "string"
                                                },
                                                "estado": {
                                                    "type": "string"
                                                },
                                                "cep": {
                                                    "type": "string"
                                                },
                                                "telefone": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            },
                            "menus": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string"
                                        },
                                        "icone": {
                                            "type": "string"
                                        },
                                        "link": {
                                            "type": "string"
                                        },
                                        "submenus": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nome": {
                                                        "type": "string"
                                                    },
                                                    "link": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/usuarios/login": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Login de usuario",
                "description": "Login de usuario",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/usuarioLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Produto criado com sucesso"
                    },
                    "400": {
                        "description": "Parâmetros inválidos"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/usuarios/cadastro": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Criar usuario",
                "description": "Criar usuario",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/usuarioCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Usuario criado com sucesso"
                    },
                    "400": {
                        "description": "Parâmetros inválidos"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/usuarios/validartoken": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Validar token JWT",
                "description": "Valida o token JWT enviado no header Authorization e retorna os dados do usuário autenticado.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/usuarioAutenticado"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    }
}